initial for stm32 port

+++ src/driver/stm32f1/delay.h
@@ -4,6 +4,6 @@
 #include <stdint.h>
 
 void delay_ms(uint16_t ms);
-void delay_us(uint16_t us);
+void delay_us(uint32_t us);
 
 #endif
